Reorganise and add authentication for update and delete requests on a given post, match logged in user to requests via oath and JWT tokens

+++ routers/posts_router.py
@@ -12,40 +12,62 @@ router = APIRouter(
 )
 
 # GET request to return all records in the Post table
-@router.get('/' , response_model=list[models.PostShow])
+@router.get('/' , response_model=list[models.PostShow], response_model_by_alias=False)
 def get_all_posts(db : Session = Depends(database.get_db)):
     posts = db.query(schemas.Post).all()
     return posts
 
 
 # Get request to return a specfic record based on id
-@router.get('/{post_id}' ,name = 'Get post by id', response_model=models.PostShow)
+@router.get('/{post_id}' ,name = 'Get post by id',
+             response_model=models.PostShow, response_model_by_alias=False)
 def get_post(post_id: int, db : Session = Depends(database.get_db)):
     post_query = postsq.get_post_query_by_id(post_id, db)
-    return post_query.first()
+    post = post_query.first()
+    return post
 
 
 # POST request to create a new record of post in the Post table
 # uses PostCreate as request body structure, Depends returns a pydantic object
-@router.post('/' , status_code=status.HTTP_201_CREATED, response_model= models.PostShow)
+@router.post('/' , status_code=status.HTTP_201_CREATED,
+              response_model= models.PostShow, response_model_by_alias=False)
 def create_post(request : models.PostCreate, db :Session = Depends(database.get_db),
                 current_user : models.TokenData = Depends(oauth2.get_current_user_id)):
     return postsq.create_post(request, db , current_user.id)
 
 
 @router.delete('/{post_id}' , status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(post_id : int , db : Session = Depends(database.get_db)):
+def delete_post(post_id : int ,
+                db : Session = Depends(database.get_db),
+                current_user : models.TokenData = Depends(oauth2.get_current_user_id)):
+    
     post_query = postsq.get_post_query_by_id(post_id , db)
-    table_name = post_query.first().__tablename__
+    post = post_query.first()
+
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail= f'Not authroised to perform the requested action')
+    
+    table_name = post.__tablename__
     num_rows = post_query.delete(synchronize_session=False)
     print(f'Deleted {num_rows} rows from {table_name.upper()} table')
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
     
-@router.put('/{post_id}' , status_code=status.HTTP_202_ACCEPTED, response_model=models.PostShow)
-def update_post(post_id : int, request : models.PostCreate, db : Session = Depends(database.get_db)):
+
+@router.put('/{post_id}' , status_code=status.HTTP_202_ACCEPTED,
+             response_model=models.PostShow, response_model_by_alias=False)
+def update_post(post_id : int, request : models.PostCreate,
+                db : Session = Depends(database.get_db),
+                current_user : models.TokenData = Depends(oauth2.get_current_user_id)):
+    
     post_query = postsq.get_post_query_by_id(post_id , db)
     post = post_query.first()
+
+    if post.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail= f'Not authroised to perform the requested action')
+    
     table_name = post.__tablename__
     num_rows = post_query.update(request.model_dump(exclude_unset=True), synchronize_session=False)
     print(f'Updated {num_rows} rows from the {table_name.upper()} table.')
